Refactored deleteAllImagesByOfferId method.

+++ src/main/java/bg/softuni/myrealestateproject/service/ImageService.java
@@ -52,13 +52,11 @@ public class ImageService {
     }
 
 
-    public boolean hasImageByOfferId(Long offerId) {
-        return this.imageRepository.existsByOfferId(offerId);
-    }
-
     public void deleteAllImagesByOfferId(Long id) {
         List<ImageEntity> allImagesWithOfferId = this.imageRepository.getAllImagesWithOfferId(id);
-        this.imageRepository.deleteAll(allImagesWithOfferId);
+        if (allImagesWithOfferId.size() > 0) {
+            this.imageRepository.deleteAll(allImagesWithOfferId);
+        }
     }
